0.0.6 - Item naming Shulker box item names now include both colors!
Boxes with a single color use the default (translatable) name.
Boxes with two colors are named:
  [primary color] And [secondary color] Shulker Box
Or, if the primary color is the normal/blank shulker box color:
  Partly [secondary color] Shulker Box

Examples:
  White And Cyan Shulker Box
  Partly Black Shulker Box

The longest combination is light blue and light gray, being 37 characters long!
  Light Blue And Light Gray Shulker Box

A system for translating the names will most likely be introduced at a later point :)

Small notes:
  - The names are currently never cached anywhere, so they're generated on every call to BlockItem.getName() on a shulker box.
  - The dual shulker color function type has been moved to DualShulkerColorLootFunction, the class for the function. No need to have a whole class for just one field