Beta 0.1 - Crafting
Beta!
The mod is now in beta. It should have been in alpha and dev versions before, but oh well. The main thing remaining for a release version is play testing and testing mod compatibility.

Crafting!
That's right, the last remaining part for survival use is now finished!

1.19.3!
The mod is updated to this version now, I guess

Also:
  - Items will continue to be rendered by reading nbt, as Minecraft already does this with a few items
  - The game doesn't crash on launch now (minor inconvenience)
  - Shulker box items with a blank top and colored bottom can now be washed in a cauldron
  - Shulkers with a colored top and blank bottom are now called "Mostly [Top Color] Shulker Box"
  - Shulker box items secondary color is now stored directly in the nbt instead of under BlockEntityTag