1.0.0 beta 5 - Particles In some situations particles from a shulker box will use the bottom or both of its colors. These situations are:   - running or landing on the side or underside of a shulker box   - breaking a shulker box with a hand, piston or command
This implementation is not perfect: when breaking a shulker box with two colors double the amount of normal particles are spawned, and landing on a sideways shulker randomly uses the top or bottom color.

Some particle effects don't work on vanilla clients connected to a modded server, so the mixins for these are client side only.

Nbt key for shulkers secondary color changed from "BaseColor" to "DualColor"

Missing from last commit:
Changed Minecraft version to 1.19.4 only